+++ tests/fixtures/_pydantic_models.py
@@ -1,12 +1,13 @@
 from __future__ import annotations
 
-from pydantic import BaseModel
+from pydantic import BaseModel, PrivateAttr
 
 
 class PydanticAddress(BaseModel):
     street: str
     house_number: int
     city: str
+    _country: str = PrivateAttr()
 
 
 class PydanticAddressIncorrectType(BaseModel):